VPAAMP-470: [Timeshift DAI] Confirm TimedMetadata Time relative to Stream
Reason for change:Added a EAP check and used the starttime of the periods
Test Procedure: Refer jira ticket VPAAMP-470
Priority: P2

+++ fragmentcollector_mpd.h
@@ -750,7 +750,7 @@ protected:
 	 * @param isInit true if its the first playlist download
 	 * @param reportBulkMeta true if bulk metadata is enabled
 	 */
-	void ProcessPeriodSupplementalProperty(Node* node, std::string& AdID, uint64_t startMS, uint64_t durationMS, bool isInit, bool reportBulkMeta=false);
+	void ProcessPeriodSupplementalProperty(Node* node, std::string& AdID, long long startMS, long long durationMS, bool isInit, bool reportBulkMeta=false);
 	/**
 	 * @fn ProcessPeriodAssetIdentifier
 	 * @param node AssetIdentifier node
@@ -761,7 +761,7 @@ protected:
 	 * @param isInit true if its the first playlist download
 	 * @param reportBulkMeta true if bulk metadata is enabled
 	 */
-	void ProcessPeriodAssetIdentifier(Node* node, uint64_t startMS, uint64_t durationMS, std::string& assetID, std::string& providerID,bool isInit, bool reportBulkMeta=false);
+	void ProcessPeriodAssetIdentifier(Node* node, long long startMS, long long durationMS, std::string& assetID, std::string& providerID,bool isInit, bool reportBulkMeta=false);
 	/**
 	 * @fn ProcessEventStream
 	 * @param startMS the start time of the event derived from the (previous) period info (ms)
@@ -769,7 +769,7 @@ protected:
 	 * @param[in] period instance.
 	 * @param reportBulkMeta true if bulk metadata is enabled
 	 */
-	bool ProcessEventStream(uint64_t startMS, int64_t startOffsetMS, IPeriod * period, bool reportBulkMeta);
+	bool ProcessEventStream(long long startMS, int64_t startOffsetMS, IPeriod * period, bool reportBulkMeta);
 	/**
 	 * @fn ProcessStreamRestrictionList
 	 * @param node StreamRestrictionListType node
